// utils/generated/exchange/subgraph.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export interface Scalars {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BigDecimal: any;
  BigInt: any;
  Bytes: any;
}

export interface Block_Height {
  readonly hash?: Maybe<Scalars["Bytes"]>;
  readonly number?: Maybe<Scalars["Int"]>;
}

export interface Bundle {
  readonly __typename?: "Bundle";
  readonly ethPrice: Scalars["BigDecimal"];
  readonly id: Scalars["ID"];
}
// ...
export interface Burn {
  readonly __typename?: "Burn";
  readonly amount0?: Maybe<Scalars["BigDecimal"]>;
  readonly amount1?: Maybe<Scalars["BigDecimal"]>;
  readonly amountUSD?: Maybe<Scalars["BigDecimal"]>;
  readonly feeLiquidity?: Maybe<Scalars["BigDecimal"]>;
  readonly feeTo?: Maybe<Scalars["Bytes"]>;
  readonly id: Scalars["ID"];
  readonly liquidity: Scalars["BigDecimal"];
  readonly logIndex?: Maybe<Scalars["BigInt"]>;
  readonly pair: Pair;
  readonly sender?: Maybe<Scalars["Bytes"]>;
  readonly timestamp: Scalars["BigInt"];
  readonly to?: Maybe<Scalars["Bytes"]>;
  readonly transaction: Transaction;
}
// ...
export interface LiquidityPosition {
  readonly __typename?: "LiquidityPosition";
  readonly id: Scalars["ID"];
  readonly liquidityTokenBalance: Scalars["BigDecimal"];
  readonly pair: Pair;
  readonly poolOwnership?: Maybe<Scalars["BigDecimal"]>;
  readonly user: User;
}
// ...
export interface Mint {
  readonly __typename?: "Mint";
  readonly amount0?: Maybe<Scalars["BigDecimal"]>;
  readonly amount1?: Maybe<Scalars["BigDecimal"]>;
  readonly amountUSD?: Maybe<Scalars["BigDecimal"]>;
  readonly feeLiquidity?: Maybe<Scalars["BigDecimal"]>;
  readonly feeTo?: Maybe<Scalars["Bytes"]>;
  readonly id: Scalars["ID"];
  readonly liquidity: Scalars["BigDecimal"];
  readonly logIndex?: Maybe<Scalars["BigInt"]>;
  readonly pair: Pair;
  readonly sender?: Maybe<Scalars["Bytes"]>;
  readonly timestamp: Scalars["BigInt"];
  readonly to: Scalars["Bytes"];
  readonly transaction: Transaction;
}
// ...
export interface Pair {
  readonly __typename?: "Pair";
  readonly burns?: Maybe<ReadonlyArray<Burn>>;
  readonly createdAtBlockNumber: Scalars["BigInt"];
  readonly createdAtTimestamp: Scalars["BigInt"];
  readonly factory: UniswapFactory;
  readonly id: Scalars["ID"];
  readonly liquidityPositions?: Maybe<ReadonlyArray<LiquidityPosition>>;
  readonly mints?: Maybe<ReadonlyArray<Mint>>;
  readonly reserve0: Scalars["BigDecimal"];
  readonly reserve1: Scalars["BigDecimal"];
  readonly reserveETH: Scalars["BigDecimal"];
  readonly reserveUSD: Scalars["BigDecimal"];
  readonly swaps?: Maybe<ReadonlyArray<Swap>>;
  readonly token0: Token;
  readonly token0Price: Scalars["BigDecimal"];
  readonly token1: Token;
  readonly token1Price: Scalars["BigDecimal"];
  readonly totalSupply: Scalars["BigDecimal"];
  readonly trackedReserveETH: Scalars["BigDecimal"];
  readonly txCount: Scalars["BigInt"];
  readonly volumeToken0: Scalars["BigDecimal"];
  readonly volumeToken1: Scalars["BigDecimal"];
  readonly volumeUSD: Scalars["BigDecimal"];
}
// ...
export interface PairDayData {
  readonly __typename?: "PairDayData";
  readonly dailyTxns: Scalars["BigInt"];
  readonly dailyVolumeToken0: Scalars["BigDecimal"];
  readonly dailyVolumeToken1: Scalars["BigDecimal"];
  readonly dailyVolumeUSD: Scalars["BigDecimal"];
  readonly date: Scalars["Int"];
  readonly id: Scalars["ID"];
  readonly pairAddress: Scalars["Bytes"];
  readonly reserve0: Scalars["BigDecimal"];
  readonly reserve1: Scalars["BigDecimal"];
  readonly reserveUSD: Scalars["BigDecimal"];
  readonly token0: Token;
  readonly token1: Token;
}
// ...
export interface Query {
  readonly __typename?: "Query";
  readonly bundle?: Maybe<Bundle>;
  readonly bundles: ReadonlyArray<Bundle>;
  readonly burn?: Maybe<Burn>;
  readonly burns: ReadonlyArray<Burn>;
  readonly liquidityPosition?: Maybe<LiquidityPosition>;
  readonly liquidityPositions: ReadonlyArray<LiquidityPosition>;
  readonly mint?: Maybe<Mint>;
  readonly mints: ReadonlyArray<Mint>;
  readonly pair?: Maybe<Pair>;
  readonly pairDayData?: Maybe<PairDayData>;
  readonly pairDayDatas: ReadonlyArray<PairDayData>;
  readonly pairs: ReadonlyArray<Pair>;
  readonly swap?: Maybe<Swap>;
  readonly swaps: ReadonlyArray<Swap>;
  readonly token?: Maybe<Token>;
  readonly tokenDayData?: Maybe<TokenDayData>;
  readonly tokenDayDatas: ReadonlyArray<TokenDayData>;
  readonly tokens: ReadonlyArray<Token>;
  readonly transaction?: Maybe<Transaction>;
  readonly transactions: ReadonlyArray<Transaction>;
  readonly uniswapDayData?: Maybe<UniswapDayData>;
  readonly uniswapDayDatas: ReadonlyArray<UniswapDayData>;
  readonly uniswapFactories: ReadonlyArray<UniswapFactory>;
  readonly uniswapFactory?: Maybe<UniswapFactory>;
  readonly user?: Maybe<User>;
  readonly users: ReadonlyArray<User>;
}
// ...
export interface Swap {
  readonly __typename?: "Swap";
  readonly amount0In: Scalars["BigDecimal"];
  readonly amount0Out: Scalars["BigDecimal"];
  readonly amount1In: Scalars["BigDecimal"];
  readonly amount1Out: Scalars["BigDecimal"];
  readonly amountUSD: Scalars["BigDecimal"];
  readonly id: Scalars["ID"];
  readonly logIndex?: Maybe<Scalars["BigInt"]>;
  readonly pair: Pair;
  readonly sender: Scalars["Bytes"];
  readonly timestamp: Scalars["BigInt"];
  readonly to: Scalars["Bytes"];
  readonly transaction: Transaction;
}
// ...
export interface Token {
  readonly __typename?: "Token";
  readonly allPairs?: Maybe<ReadonlyArray<Pair>>;
  readonly decimals: Scalars["BigInt"];
  readonly derivedETH?: Maybe<Scalars["BigDecimal"]>;
  readonly id: Scalars["ID"];
  readonly mostLiquidPairs: ReadonlyArray<Maybe<PairDayData>>;
  readonly name: Scalars["String"];
  readonly symbol: Scalars["String"];
  readonly totalLiquidity: Scalars["BigDecimal"];
  readonly tradeVolume: Scalars["BigDecimal"];
  readonly tradeVolumeUSD: Scalars["BigDecimal"];
  readonly txCount: Scalars["BigInt"];
}
// ...
export interface TokenDayData {
  readonly __typename?: "TokenDayData";
  readonly dailyTxns: Scalars["BigInt"];
  readonly dailyVolumeETH: Scalars["BigDecimal"];
  readonly dailyVolumeToken: Scalars["BigDecimal"];
  readonly dailyVolumeUSD: Scalars["BigDecimal"];
  readonly date: Scalars["Int"];
  readonly id: Scalars["ID"];
  readonly maxStored: Scalars["Int"];
  readonly mostLiquidPairs: ReadonlyArray<PairDayData>;
  readonly priceUSD: Scalars["BigDecimal"];
  readonly token: Token;
  readonly totalLiquidityETH: Scalars["BigDecimal"];
  readonly totalLiquidityToken: Scalars["BigDecimal"];
  readonly totalLiquidityUSD: Scalars["BigDecimal"];
}
// ...
export interface Transaction {
  readonly __typename?: "Transaction";
  readonly blockNumber: Scalars["BigInt"];
  readonly burns: ReadonlyArray<Maybe<Burn>>;
  readonly id: Scalars["ID"];
  readonly mints: ReadonlyArray<Maybe<Mint>>;
  readonly swaps: ReadonlyArray<Maybe<Swap>>;
  readonly timestamp: Scalars["BigInt"];
}
// ...
export interface UniswapDayData {
  readonly __typename?: "UniswapDayData";
  readonly dailyVolumeETH: Scalars["BigDecimal"];
  readonly dailyVolumeUSD: Scalars["BigDecimal"];
  readonly date: Scalars["Int"];
  readonly id: Scalars["ID"];
  readonly maxStored?: Maybe<Scalars["Int"]>;
  readonly mostLiquidTokens: ReadonlyArray<TokenDayData>;
  readonly totalLiquidityETH: Scalars["BigDecimal"];
  readonly totalLiquidityUSD: Scalars["BigDecimal"];
  readonly totalVolumeETH: Scalars["BigDecimal"];
  readonly totalVolumeUSD: Scalars["BigDecimal"];
  readonly txCount: Scalars["BigInt"];
}
// ...
export interface UniswapFactory {
  readonly __typename?: "UniswapFactory";
  readonly id: Scalars["ID"];
  readonly mostLiquidTokens: ReadonlyArray<TokenDayData>;
  readonly pairCount: Scalars["Int"];
  readonly pairs: ReadonlyArray<Maybe<Pair>>;
  readonly totalLiquidityETH: Scalars["BigDecimal"];
  readonly totalLiquidityUSD: Scalars["BigDecimal"];
  readonly totalVolumeETH: Scalars["BigDecimal"];
  readonly totalVolumeUSD: Scalars["BigDecimal"];
  readonly untrackedVolumeUSD: Scalars["BigDecimal"];
  readonly txCount: Scalars["BigInt"];
}
// ...
export interface User {
  readonly __typename?: "User";
  readonly id: Scalars["ID"];
  readonly liquidityPositions?: Maybe<ReadonlyArray<LiquidityPosition>>;
}
// ...
export type UniswapFactoriesQueryVariables = Exact<{
  blockNumber: Scalars["Int"];
  factoryAddress: Scalars["ID"];
}>;

export type UniswapFactoriesQuery = { readonly __typename?: "Query" } & {
  readonly uniswapFactories: ReadonlyArray<
    { readonly __typename?: "UniswapFactory" } & Pick<
      UniswapFactory,
      | "id"
      | "totalVolumeUSD"
      | "totalVolumeETH"
      | "untrackedVolumeUSD"
      | "totalLiquidityUSD"
      | "totalLiquidityETH"
      | "txCount"
      | "pairCount"
    >
  >;
};

export type CurrentUniswapFactoriesQueryVariables = Exact<{
  factoryAddress: Scalars["ID"];
}>;

export type CurrentUniswapFactoriesQuery = { readonly __typename?: "Query" } & {
  readonly uniswapFactories: ReadonlyArray<
    { readonly __typename?: "UniswapFactory" } & Pick<
      UniswapFactory,
      | "id"
      | "totalVolumeUSD"
      | "totalVolumeETH"
      | "untrackedVolumeUSD"
      | "totalLiquidityUSD"
      | "totalLiquidityETH"
      | "txCount"
      | "pairCount"
    >
  >;
};